+++ RELEASE_NOTES.md
@@ -1,7 +1,12 @@
+### Version 0.4 (February 27, 2014)
+
+* Correctly capture command line output for external programs on Windows.
+
 ### Version 0.3 (February 26, 2014)
 
 * Fix how the Vagrant executable is run on Windows - [Issue 1](https://github.com/bmuschko/gradle-vagrant-plugin/issues/1).
 * Default to `projectDir` for `boxDir` property - [Issue 2](https://github.com/bmuschko/gradle-vagrant-plugin/issues/2).
+* Upgrade to Gradle Wrapper 1.11.
 
 ### Version 0.2 (September 28, 2013)